+++ tests/test_main.py
@@ -419,6 +419,19 @@ class TestNamespaces(FabricTest):
             eq_(len(funcs), 1)
             ok_('foo' in funcs)
 
+    def test_class_based_tasks_are_found_with_variable_name(self):
+        """
+        A new-style tasks with undefined name attribute should use the instance
+        variable name.
+        """
+        module = fabfile('classbased_task_fabfile.py')
+        from fabric.state import env
+        with path_prefix(module):
+            docs, funcs = load_fabfile(module)
+            eq_(len(funcs), 1)
+            ok_('foo' in funcs)
+            eq_(funcs['foo'].name, 'foo')
+
     def test_recursion_steps_into_nontask_modules(self):
         """
         Recursive loading will continue through modules with no tasks